refactor: dissolve rpc/evo_util.cpp into rpc/evo.cpp
Also drop coinjoin/client.cpp's unused include of the header and correct the
stale note claiming CDeterministicMN::ToJson() lived in
evo/deterministicmns.cpp: it lived in rpc/evo_util.cpp and now lives in
rpc/evo.cpp.
